flask-course/requirements.txt
alembic==1.4.2
click==7.1.2
flake8==3.8.3
Flask==1.1.2
Flask-Env==2.0.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
